Remove old AllVenture report
- Remove old AllVenture report
