+++ src/ralph_pricing/tests/collect_plugins/test_ceilometer.py
