+++ src/ralph_pricing/views/base.py
@@ -24,12 +24,6 @@ MAIN_MENU = [
         view_name='all_ventures',
     ),
     MenuItem(
-        _("All ventures BETA"),
-        name='all-ventures-beta',
-        fugue_icon='fugue-store-medium',
-        view_name='all_ventures_beta',
-    ),
-    MenuItem(
         _("Devices"),
         name='devices',
         fugue_icon='fugue-wooden-box',
